Add PageRankCommitsGraph

+++ core/src/main/kotlin/visualisation/entity/NodeThreeJS.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class NodeThreeJS(
     val id: String,
-    val value: Int = 1,
+    val value: Float = 1f,
     val color: String = "#323ca8",
     val shape: Int = 0
 ) : Comparable<NodeThreeJS> {
